Fix round manager, remove shop for now and misc fixes
