+++ mod/TTT.Round/RoundServiceExtension.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using TTT.Public.Extensions;
+using TTT.Public.Mod.Round;
+
+namespace TTT.Round;
+
+public static class RoundServiceExtension
+{
+    public static void AddRoundService(this IServiceCollection collection)
+    {
+        collection.AddPluginBehavior<IRoundService, RoundBehavior>();
+    }
+}
