--- mod/TTT.Shop/Items/Traitor/AwpItem.cs
@@ -1,33 +1,0 @@
-﻿using CounterStrikeSharp.API.Modules.Entities.Constants;
-using TTT.Player;
-using TTT.Public.Mod.Role;
-using TTT.Public.Shop;
-
-namespace TTT.Shop.Items.Traitor;
-
-public class AwpItem : IShopItem
-{
-    public string Name()
-    {
-        return "AWP";
-    }
-
-    public string SimpleName()
-    {
-        return "awp";
-    }
-
-    public int Price()
-    {
-        return 2000;
-    }
-
-    public BuyResult OnBuy(GamePlayer player)
-    {
-        if (player.Credits() < Price()) return BuyResult.NotEnoughCredits;
-        if (player.PlayerRole() != Role.Detective) return BuyResult.IncorrectRole;
-        player.RemoveCredits(Price());
-        player.Player().GiveNamedItem(CsItem.AWP);
-        return BuyResult.Successful;
-    }
-}
