--- mod/TTT.Shop/TraitorShopHandler.cs
@@ -1,22 +1,0 @@
-﻿
-
-using TTT.Shop.Items;
-
-namespace TTT.Shop;
-
-public class TraitorShopHandler : BaseShopHandler
-{
- 
-    private static readonly TraitorShopHandler _instance = new TraitorShopHandler();
-    
-    public TraitorShopHandler()
-    {
-        AddItems("Traitor");
-    }
-    
-    public new static TraitorShopHandler Get()
-    {
-        return _instance;
-    }
-    
-}
