--- public/TTT.Public/Shop/BuyResult.cs
@@ -1,9 +1,0 @@
-﻿namespace TTT.Public.Shop;
-
-public enum BuyResult
-{
-    Successful,
-    NotEnoughCredits,
-    AlreadyOwned,
-    IncorrectRole
-}
