--- public/TTT.Public/Shop/IActionItem.cs
@@ -1,8 +1,0 @@
-﻿using TTT.Player;
-
-namespace TTT.Public.Shop;
-
-public interface IActionItem : IShopItem
-{
-    void OnAction(GamePlayer player);
-}
