--- public/TTT.Public/Shop/IShopItemHandler.cs
@@ -1,7 +1,0 @@
-﻿namespace TTT.Public.Shop;
-
-public interface IShopItemHandler
-{
-    ISet<IShopItem> GetShopItems();
-    void AddShopItem(IShopItem item);
-}
