+++ public/TTT.Public/TTT.Public.csproj
@@ -8,7 +8,4 @@
     <PackageReference Include="CounterStrikeSharp.API" Version="1.0.247" />
     <PackageReference Include="MySqlConnector" Version="2.3.6" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Shop\Items\" />
-  </ItemGroup>
 </Project>
